+++ src/domain/devices/low_pass_filter_effect.hpp
@@ -38,14 +38,14 @@ public:
         return typeIdString();
     }
 
-    void setCutoff(float cutoff);
-    void process(float & left, float & right) override;
+    void setCutoff(double cutoff);
+    void process(double & left, double & right) override;
     void process(AudioContext & context) override;
     void reset() override;
 
 private:
-    void processSample(float & left, float & right, double g, double damping, double k);
-    float m_cutoff { 1.0f };
+    void processSample(double & left, double & right, double g, double damping, double k);
+    double m_cutoff { 1.0 };
     double m_s1L { 0.0 }, m_s2L { 0.0 };
     double m_s1R { 0.0 }, m_s2R { 0.0 };
 };
