+++ src/domain/dsp/svf_filter.hpp
@@ -41,7 +41,7 @@ public:
 
     void setBypass();
 
-    float process(float input);
+    double process(double input);
     void reset();
 
 private:
